\name{leafdev}
\docType{data}
\alias{leafdev}
\title{Example data}
\description{Example data set (\code{TilingFeatureSet}) on leaf development in the plant Arabidopsis thaliana [1] for use of the waveTiling package. The data set contains 6553600 features and 18 samples. Transcriptome analysis was conducted for 6 developmental time points (day 8 to day 13), with 3 biological replicates per time point. The focus of the initial study was to unravel the underlying mechanisms of on one hand the transition from cell division to cell expansion and on the other hand the transition from non-photosynthetic to photosynthetic leaves.}
\usage{data(leafdev)}
\references{
[1] Andriankaja M, Dhondt S, De Bodt S, Vanhaeren H, Coppens F, et al. (2012) Exit from proliferation during leaf development in Arabidopsis thaliana: A not-so-gradual process. Developmental Cell 22: 64-78.}
\keyword{datasets}
\examples{
  data(leafdev)
}
