\name{leafdevBQ}
\docType{data}
\alias{leafdevBQ}
\title{Example data}
\description{Example data set (\code{TilingFeatureSet}) on leaf development in the plant Arabidopsis thaliana for use of the waveTiling package. The data are taken from [1]. The dataset contains the background-corrected and quantile-normalized expression data}
\usage{data(leafdevBQ)}
\references{
[1] Andriankaja M, Dhondt S, De Bodt S, Vanhaeren H, Coppens F, et al. (2012) Exit from proliferation during leaf development in Arabidopsis thaliana: A not-so-gradual process. Developmental Cell 22: 64-78.}
\keyword{datasets}
\examples{
  data(leafdevBQ)
}
