\name{leafdevFit}
\docType{data}
\alias{leafdevFit}
\title{Example waveTiling fit object}
\description{Example \code{WfmFit}-class object as output after fitting the wavelet-based functional model to the \code{leafdev} data for the forward strand of chromosome 1.}
\usage{data(leafdevFit)}
\references{ }
\keyword{datasets}
\examples{
  data(leafdevFit)
}
