\name{leafdevInfCompare}
\docType{data}
\alias{leafdevInfCompare}
\title{Example waveTiling inference object}
\description{Example \code{WfmInf}-class object as output after transcriptome analysis of the \code{leafdev} data for the forward strand of chromosome 1, using pairwise comparisons between the different time points.}
\usage{data(leafdevInfCompare)}
\references{ }
\keyword{datasets}
\examples{
  data(leafdevInfCompare)
}
