\name{leafdevMapAndFilterTAIR9}
\docType{data}
\alias{leafdevMapAndFilterTAIR9}
\title{Example waveTiling \code{mapFilterProbe} object}
\description{Example \code{mapFilterProbe}-class object as output after filtering redundant probes (PM/MM and/or forward/reverse strand) and remapping the probes to the Arabidopsis thaliana TAIR9 genome sequence.}
\usage{data(leafdevMapAndFilterTAIR9)}
\references{ }
\keyword{datasets}
\examples{
  data(leafdevMapAndFilterTAIR9)
}
